+++ skbio/sequence/_dna.py
@@ -7,15 +7,17 @@
 # ----------------------------------------------------------------------------
 
 from __future__ import absolute_import, division, print_function
+from six import add_metaclass
 
 import skbio
 from skbio.util._decorator import classproperty, overrides
 from skbio.util._decorator import stable
 from ._nucleotide_mixin import NucleotideMixin, _motifs as _parent_motifs
-from ._iupac_sequence import IUPACSequence
+from ._grammared_sequence import GrammaredSequence, DisableSubclassingMeta
 
 
-class DNA(IUPACSequence, NucleotideMixin):
+@add_metaclass(DisableSubclassingMeta)
+class DNA(GrammaredSequence, NucleotideMixin):
     """Store DNA sequence data and optional associated metadata.
 
     Only characters in the IUPAC DNA character set [1]_ are supported.
@@ -64,6 +66,14 @@ class DNA(IUPACSequence, NucleotideMixin):
     See Also
     --------
     RNA
+    GrammaredSequence
+
+    Notes
+    -----
+    Subclassing is disabled for DNA, because subclassing makes
+    it possible to change the alphabet, and certain methods rely on the
+    IUPAC alphabet. If a custom sequence alphabet is needed, inherit directly
+    from ``GrammaredSequence``.
 
     References
     ----------
@@ -118,19 +128,48 @@ class DNA(IUPACSequence, NucleotideMixin):
 
     @classproperty
     @stable(as_of="0.4.0")
-    @overrides(IUPACSequence)
+    @overrides(GrammaredSequence)
     def nondegenerate_chars(cls):
         return set("ACGT")
 
     @classproperty
     @stable(as_of="0.4.0")
-    @overrides(IUPACSequence)
+    @overrides(GrammaredSequence)
     def degenerate_map(cls):
         return {
             "R": set("AG"), "Y": set("CT"), "M": set("AC"), "K": set("TG"),
             "W": set("AT"), "S": set("GC"), "B": set("CGT"), "D": set("AGT"),
             "H": set("ACT"), "V": set("ACG"), "N": set("ACGT")
         }
+
+    @classproperty
+    def default_gap_char(cls):
+        """Gap character to use when constructing a new gapped sequence.
+
+        This character is used when it is necessary to represent gap characters
+        in a new sequence. For example, a majority consensus sequence will use
+        this character to represent gaps.
+
+        Returns
+        -------
+        str
+            Default gap character.
+
+        """
+        return '-'
+
+    @classproperty
+    @stable(as_of='0.4.0')
+    def gap_chars(cls):
+        """Return characters defined as gaps.
+
+        Returns
+        -------
+        set
+            Characters defined as gaps.
+
+        """
+        return set('-.')
 
     @property
     def _motifs(self):
@@ -399,7 +438,7 @@ class DNA(IUPACSequence, NucleotideMixin):
         """
         return self.transcribe().translate_six_frames(*args, **kwargs)
 
-    @overrides(IUPACSequence)
+    @overrides(GrammaredSequence)
     def _repr_stats(self):
         """Define custom statistics to display in the sequence's repr."""
         stats = super(DNA, self)._repr_stats()